refactor(identity): move mangled-name-genuine determination to model layer
ADR-063 Phase 2, sixth slice: relocate is_real_mangled_name/
normalize_mangled_name and the ~450-line Itanium-mangling-validation
machinery behind them out of finding_identity.py (comparison-layer
code) into a new leaf module, model/mangled_name_validity.py.

model/identity.py's entity_id_for_function/entity_id_for_variable and
finding_identity.py's resolve_function_identity/resolve_variable_identity
both need this same determination; moving the algorithm down to a shared
leaf primitive (rather than finding_identity.py owning it, which would
create a forbidden leaf -> comparison-layer import) lets both sides call
one implementation instead of keeping independent copies.

model/identity.py and finding_identity.py both re-export
is_real_mangled_name/normalize_mangled_name under their original names,
so every existing call site is unaffected -- no behavior change.

Named mangled_name_validity.py, not mangled_name.py: that name was
already taken by ADR-061 D1's unrelated Itanium scope-component parsing
chain. Tests moved to tests/test_mangled_name_validity.py for the same
reason (tests/test_mangled_name.py already existed for that other
module).

Updates the ADR-063 plan doc and ADR status paragraph to record this
slice; the two remaining Phase 2 items (unifying finding_identity's
identity shape with EntityId, and the post-parse consumer migrations
in diff_filtering.py/type_reachability.py) stay open.
