+++ changelog.d/20260830_170523_noreply_adr_063_oa3f5x.md
@@ -1,0 +1,13 @@
+### Changed
+
+- **The mangled-name-is-genuine determination (`is_real_mangled_name`/
+  `normalize_mangled_name`) moved from `finding_identity.py` into a new
+  leaf module, `model/mangled_name_validity.py`** (ADR-063 Phase 2, sixth
+  slice). `finding_identity.py` now re-exports both names as thin
+  delegating imports instead of owning the ~450-line, independently
+  multi-round-reviewed Itanium-mangling-validation algorithm — the same
+  determination `model.identity.entity_id_for_function`/
+  `entity_id_for_variable` need is now answered by one shared primitive
+  instead of an independent copy. No behavior change; every existing
+  `finding_identity.is_real_mangled_name`/`normalize_mangled_name` call
+  site is unaffected.
